src/python/requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
python-magic>=0.4.27
unstructured>=0.10.30
pyPDF2>=3.0.0
chromadb>=0.4.22
gradio>=4.0.0
moviepy>=1.0.3
SpeechRecognition>=3.10.0
langchain_community
sentence_transformers
langchain_huggingface
openai-whisper
